+++ src/app/(settings)/setting/page.tsx
@@ -1,0 +1,13 @@
+import GPSGroup from "../_components/GPSGroup";
+import SignOutButton from "../_components/SignOutButtont";
+
+const SettingPage = async () => {
+  return (
+    <div>
+      <GPSGroup />
+      <SignOutButton />
+    </div>
+  );
+};
+
+export default SettingPage;
